requirements.txt
feedparser==6.0.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sendgrid==6.11.0
openai==1.47.0
httpx==0.27.2
